Headstorm_AI/blueprint_analysis/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
numpy>=1.23.0
Pillow>=9.5.0
easyocr>=1.6.2
gradio>=3.35.0
roboflow>=1.1.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
pandas>=2.0.0
